@@ -11,37 +11,55 @@
     "lint": "ionic-app-scripts lint"
   },
   "dependencies": {
-    "@angular/animations": "5.2.11",
-    "@angular/common": "5.2.11",
-    "@angular/compiler": "5.2.11",
+    "@angular/animations": "6.0.9",
+    "@angular/cdk": "^6.3.3",
+    "@angular/common": "6.0.9",
+    "@angular/compiler": "6.0.9",
     "@angular/compiler-cli": "5.2.11",
-    "@angular/core": "5.2.11",
-    "@angular/forms": "5.2.11",
+    "@angular/core": "6.0.9",
+    "@angular/forms": "6.0.9",
     "@angular/http": "5.2.11",
+    "@angular/material": "^6.3.3",
     "@angular/platform-browser": "5.2.11",
     "@angular/platform-browser-dynamic": "5.2.11",
     "@ionic-native/camera": "^4.10.0",
     "@ionic-native/core": "~4.10.0",
+    "@ionic-native/firebase": "^4.11.0",
+    "@ionic-native/geolocation": "^4.12.0",
     "@ionic-native/splash-screen": "^4.10.0",
     "@ionic-native/status-bar": "~4.10.0",
     "@ionic/pro": "1.0.20",
     "@ionic/storage": "2.1.3",
-    "cordova-android": "^7.0.0",
+    "@tensorflow/tfjs": "^0.12.5",
+    "angularfire2": "^5.0.0-rc.11",
+    "cordova-android": "7.1.1",
+    "cordova-browser": "^5.0.4",
     "cordova-plugin-camera": "^4.0.3",
     "cordova-plugin-device": "^2.0.2",
-    "cordova-plugin-ionic-keyboard": "^2.0.5",
-    "cordova-plugin-ionic-webview": "^1.1.19",
+    "cordova-plugin-firebase": "^1.0.5",
+    "cordova-plugin-geolocation": "^4.0.1",
+    "cordova-plugin-ionic-keyboard": "^2.1.2",
+    "cordova-plugin-ionic-webview": "^1.2.1",
     "cordova-plugin-splashscreen": "^5.0.2",
+    "cordova-plugin-statusbar": "^2.4.2",
     "cordova-plugin-whitelist": "^1.3.3",
+    "firebase": "^5.3.1",
+    "firebase-tools": "^4.1.0",
     "ionic-angular": "3.9.2",
     "ionicons": "3.0.0",
+    "observable": "^2.1.4",
     "package.json": "^2.0.1",
-    "rxjs": "5.5.11",
+    "require": "^2.4.20",
+    "rxjs": "^6.2.2",
+    "rxjs-compat": "^6.2.2",
     "sw-toolbox": "3.6.0",
+    "yarn": "^1.9.4",
     "zone.js": "0.8.26"
   },
   "devDependencies": {
+    "@angular/forms": "^5.2.5",
     "@ionic/app-scripts": "3.1.11",
+    "@types/node": "^10.7.0",
     "typescript": "~2.6.2"
   },
   "description": "An Ionic project",
@@ -52,9 +70,15 @@
       "cordova-plugin-device": {},
       "cordova-plugin-ionic-webview": {},
       "cordova-plugin-ionic-keyboard": {},
-      "cordova-plugin-camera": {}
+      "cordova-plugin-camera": {},
+      "cordova-plugin-statusbar": {},
+      "cordova-plugin-firebase": {},
+      "cordova-plugin-geolocation": {
+        "GEOLOCATION_USAGE_DESCRIPTION": "User location for queries"
+      }
     },
     "platforms": [
+      "browser",
       "android"
     ]
   }
