@@ -1,22 +1,54 @@
-import { Component } from '@angular/core';
-import { Platform } from 'ionic-angular';
+import { Component  , ViewChild} from '@angular/core';
+import { Nav, Platform } from 'ionic-angular';
 import { StatusBar } from '@ionic-native/status-bar';
 import { SplashScreen } from '@ionic-native/splash-screen';
-import { Camera, CameraOptions } from '@ionic-native/camera';
-import { TabsPage } from '../pages/tabs/tabs';
+// import { Camera, CameraOptions } from '@ionic-native/camera';
 
+import { AboutPage } from '../pages/about/about';
+import { ContactPage } from '../pages/contact/contact';
+// import { HomePage } from '../pages/home/home';
+import { TabsPage } from '../pages/tabs/tabs';
+// import { UtilitiesPage } from '../pages/utilities/utilities';
+import { ImagerecPage } from '../pages/imagerec/imagerec';
+import { ModelLoaderProvider } from '../providers/model-loader/model-loader';
+
+
+export interface PageInterface {
+  title: string;
+  pageName: string;
+  tabComponent?: any;
+  index?: number;
+  icon: string;
+}
 @Component({
   templateUrl: 'app.html'
 })
 export class MyApp {
+@ViewChild(Nav) nav: Nav;
   rootPage:any = TabsPage;
-
-  constructor(platform: Platform, statusBar: StatusBar, splashScreen: SplashScreen) {
+  isLoaded = false;
+  pages: Array<{ title: string, component: any }>;
+  constructor(platform: Platform, statusBar: StatusBar, splashScreen: SplashScreen, modelLoader: ModelLoaderProvider) {
     platform.ready().then(() => {
-      // Okay, so the platform is ready and our plugins are available.
-      // Here you can do any higher level native things you might need.
       statusBar.styleDefault();
-     // splashScreen.hide();
+
+      console.log('App: Calling provider function - loadModel');
+      modelLoader.loadModel();
     });
+  
+    this.pages = [
+      { title: 'Custom image recognition', component: ImagerecPage },
+      { title: 'Need Help?', component: ContactPage },
+      { title: 'About Us', component: AboutPage }
+    ];
+  }
+
+  openPage(page) {
+    this.nav.setRoot(page.component);
+  }
+  
+  isActive(page)
+  {
+
   }
 }
