@@ -1,3 +1,4 @@
+import { ModelLoaderProvider } from './../providers/model-loader/model-loader';
 import { NgModule, ErrorHandler } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { IonicApp, IonicModule, IonicErrorHandler } from 'ionic-angular';
@@ -5,19 +6,32 @@ import { MyApp } from './app.component';
 
 import { AboutPage } from '../pages/about/about';
 import { ContactPage } from '../pages/contact/contact';
+import { FeedbackPage } from './../pages/feedback/feedback';
 import { HomePage } from '../pages/home/home';
 import { TabsPage } from '../pages/tabs/tabs';
 import { UtilitiesPage } from '../pages/utilities/utilities';
 import { ImagerecPage } from '../pages/imagerec/imagerec';
+import { ResultsPage } from '../pages/results/results';
+
 
 import { StatusBar } from '@ionic-native/status-bar';
 import { SplashScreen } from '@ionic-native/splash-screen';
 import { Camera } from '@ionic-native/camera';
+import { Firebase } from '@ionic-native/firebase';
+import { Geolocation } from '@ionic-native/geolocation';
 
 import { HttpModule } from '@angular/http';
 import { HttpClientModule } from '@angular/common/http';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { FormControl } from '@angular/forms';
+import { FormsModule } from '@angular/forms';
+
+import {AngularFireModule} from 'angularfire2';
+import {AngularFireDatabaseModule /*, AngularFireDatabase,*/ } from 'angularfire2/database';
+import {AngularFireAuthModule} from 'angularfire2/auth';
+import { MatCardModule, MatTableModule } from '@angular/material';
+import { CdkTableModule } from '@angular/cdk/table';
+import { AngularFirestoreModule } from 'angularfire2/firestore';
+import { environment } from './../environments/environment';
+import { AngularFireStorageModule } from 'angularfire2/storage';
 
 @NgModule({
   declarations: [
@@ -27,14 +41,26 @@ import { FormControl } from '@angular/forms';
     HomePage,
     TabsPage,
     UtilitiesPage,
-    ImagerecPage
+    ImagerecPage,
+    FeedbackPage,
+    ResultsPage
   ],
   imports: [
     BrowserModule,
-    IonicModule.forRoot(MyApp),
+    IonicModule.forRoot(MyApp, {
+      tabsHideOnSubPages: false
+   }),
     HttpModule,
     HttpClientModule,
-    FormsModule
+    FormsModule,
+    AngularFireModule.initializeApp(environment.firebase),
+    AngularFireDatabaseModule,
+    AngularFireAuthModule,
+    MatCardModule,
+    MatTableModule,
+    CdkTableModule,
+    AngularFirestoreModule,
+    AngularFireStorageModule
   ],
   bootstrap: [IonicApp],
   entryComponents: [
@@ -44,13 +70,21 @@ import { FormControl } from '@angular/forms';
     HomePage,
     TabsPage,
     UtilitiesPage,
-    ImagerecPage
+    ImagerecPage,
+    FeedbackPage,
+    ResultsPage
   ],
   providers: [
     StatusBar,
     SplashScreen,
     Camera,
+    AngularFirestoreModule,
+    Firebase,
+    ModelLoaderProvider,
+    Geolocation,
     {provide: ErrorHandler, useClass: IonicErrorHandler}
   ]
 })
-export class AppModule {}
+export class AppModule {
+
+}
