@@ -1,6 +1,8 @@
 import { Component } from '@angular/core';
-import { IonicPage, NavController, NavParams } from 'ionic-angular';
-
+import { NavController, NavParams, AlertController} from 'ionic-angular';
+import { ModalController } from 'ionic-angular';
+// import { ContactPage } from '../contact/contact';
+import { AboutPage } from '../about/about';
 /**
  * Generated class for the UtilitiesPage page.
  *
@@ -8,7 +10,6 @@ import { IonicPage, NavController, NavParams } from 'ionic-angular';
  * Ionic pages and navigation.
  */
 
-@IonicPage()
 @Component({
   selector: 'page-utilities',
   templateUrl: 'utilities.html',
@@ -17,23 +18,50 @@ export class UtilitiesPage {
 	Math: Math = Math;
 	totalObjects : number =null;
 
-  constructor(public navCtrl: NavController, public navParams: NavParams) {
+  
+  constructor(public navCtrl: NavController, public navParams: NavParams, 
+    public alertController: AlertController, public modalCtrl : ModalController) {
 
   }
 
+  openModal()
+  {
+    var data = { message : 'hello world' };
+    var homePage = this.modalCtrl.create(AboutPage,data);
+    homePage.present();
+  }
   ionViewDidLoad() {
     console.log('ionViewDidLoad UtilitiesPage');
   }
-	  single_item = 1;
-	  empty_bucket = 1;
-	  filled_bucket = 1;
 
+  // errorAlert =function() {
+  //   let erralert = this.alertController.create({
+  //     title: 'Message sent!',
+  //     subTitle: 'Your message has been sent. A member of our team will get back to you as soon as possible.',
+  //     buttons: ['Dismiss']
+  //   })
+  // }
+
+    presentAlert =function() {
+      let alert = this.alertController.create({
+        title: 'Total amount of items',
+        subTitle: this.totalObjects,
+        buttons: ['Dismiss']
+      });
+      alert.present();
+    }
+
+    //single filed empty
   hello = function(a: number, b: number, c: number)
   {
-   		this.totalObjects = ((b - c)/a);
-   			//alert(totalObjects);
-   			return this.totalObjects;
+  
+
+      this.totalObjects = ((b - c)/a);
+      //	alert(this.totalObjects);
+      this.presentAlert();
+      return this.totalObjects;
 
    }
+   
 
 }
