@@ -1,18 +1,23 @@
 import { ContactUsComponent } from './contact-us/contact-us.component';
 import { MaterialModule } from './material.module';
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule, CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA } from '@angular/core';
+import { NgModule, NO_ERRORS_SCHEMA } from '@angular/core';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AppRoutingModule } from './/app-routing.module';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { FormControl } from '@angular/forms';
+import { AngularFireModule } from 'angularfire2';
+import { AngularFireStorageModule } from 'angularfire2/storage';
+import { BrowserXhr } from '@angular/http';
 
 import { AppComponent } from './app.component';
 import { HomeComponent } from './home/home.component';
 import { ImageuploadComponent } from './imageupload/imageupload.component';
 import { MatGridListModule } from '@angular/material';
 import { UtilitiesComponent } from './utilities/utilities.component';
+import { FeedbackComponent } from './feedback/feedback.component';
 import { HttpClientModule } from '@angular/common/http';
+import { environment } from './../environments/environment';
+import { LocationStrategy, HashLocationStrategy } from '../../node_modules/@angular/common';
 
 @NgModule({
   declarations: [
@@ -20,7 +25,8 @@ import { HttpClientModule } from '@angular/common/http';
     HomeComponent,
     ImageuploadComponent,
     UtilitiesComponent,
-    ContactUsComponent
+    ContactUsComponent,
+    FeedbackComponent
   ],
   imports: [
     BrowserModule,
@@ -30,10 +36,16 @@ import { HttpClientModule } from '@angular/common/http';
     MatGridListModule,
     FormsModule,
     HttpClientModule,
-    ReactiveFormsModule
+    ReactiveFormsModule,
+    AngularFireModule.initializeApp(environment.firebase),
+    AngularFireStorageModule
   ],
-  providers: [],
-  bootstrap: [AppComponent],
+  providers: [
+    {provide: LocationStrategy, useClass: HashLocationStrategy }
+  ],
+  bootstrap: [
+    AppComponent
+  ],
   schemas: [NO_ERRORS_SCHEMA]
 })
 export class AppModule { }
