@@ -1,8 +1,9 @@
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule, CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MatButtonModule, MatToolbarModule, MatProgressSpinnerModule, MatCardModule, MatGridListModule } from '@angular/material';
-import { MatInputModule, MatListModule, MatDividerModule } from '@angular/material';
+import { MatButtonModule, MatToolbarModule } from '@angular/material';
+import { MatInputModule, MatListModule, MatDividerModule, MatTableModule, MatSelectModule , MatCardModule} from '@angular/material';
+import { MatGridListModule, MatIconModule, MatProgressSpinnerModule  } from '@angular/material';
 
 @NgModule({
   imports: [
@@ -18,7 +19,10 @@ import { MatInputModule, MatListModule, MatDividerModule } from '@angular/materi
     MatListModule,
     MatDividerModule,
     MatGridListModule,
-    BrowserAnimationsModule
+    MatTableModule,
+    BrowserAnimationsModule,
+    MatSelectModule,
+    MatIconModule
   ],
   declarations: [],
   schemas: [NO_ERRORS_SCHEMA]
