@@ -1,3 +1,11 @@
 export const environment = {
-  production: true
+  production: true,
+  firebase: {
+    apiKey: 'AIzaSyBFmgwBqG14FM-umd-Jz1fi7pfVt2rDlAc',
+    authDomain: 'testproject-ee885.firebaseapp.com',
+    databaseURL: 'https://testproject-ee885.firebaseio.com',
+    projectId: 'testproject-ee885',
+    storageBucket: 'testproject-ee885.appspot.com',
+    messagingSenderId: '55902150213'
+  }
 };
