@@ -4,5 +4,13 @@
 // The list of which env maps to which file can be found in `.angular-cli.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  firebase: {
+    apiKey: 'AIzaSyBFmgwBqG14FM-umd-Jz1fi7pfVt2rDlAc',
+    authDomain: 'testproject-ee885.firebaseapp.com',
+    databaseURL: 'https://testproject-ee885.firebaseio.com',
+    projectId: 'testproject-ee885',
+    storageBucket: 'testproject-ee885.appspot.com',
+    messagingSenderId: '55902150213'
+  }
 };
